Refactor: eventListenerGenerator
